Recursive setequal for lists

Gets recursive setequal for lists.

set_equal(x, y, .expr = NULL, .name = NULL, .depth = 0L)

# S4 method for ANY,ANY
set_equal(x, y, .expr = FALSE, .name = NULL, .depth = 0L)

# S4 method for list,list
set_equal(x, y, .expr = NULL, .name = NULL, .depth = 0L)

Arguments

x

(list, atomic) object to compare.

y

(list, atomic) obejct to compare with.

.expr

(language) any syntactically valid R expression. Argument is parsed, evaluated and passed recursivelly to calls run on the nested elements of x and y. Other arguments (.name, .depth and .type) can be used to influence output of the comparisons. For example if (.depth == 2) return(x) will return x for the nested list elements of depth equal 2. Note: If .expr has a return statement please be aware that it should be the same as the type of the output.

.name

(character(1)) Private argument, please don't specify. Used to identify the name of the current list element when calld resursivelly. .name can be used with .expr.

.depth

(integer(1)) Private argument, please don't specify. depth of the call. Increases by 1 according to depth of the lists x and y

Value

object of the same class as x or NULL if doesn't differ from y.

Examples

set_equal(c(1, 2), c(2, 3))
#> [1] FALSE
set_equal(list(a = 1, b = 2), list(b = 2, c = 3))
#> [1] FALSE
set_equal(letters[1:2], letters[2:3])
#> [1] FALSE